+++ R/county_map.R
@@ -1,7 +1,11 @@
 #' Dataset with the lat. / long. of county FIPS codes used for mapping.
 #'
-#' Built-in dataset for use with the \code{bls_map_county} function.
+#' Built-in dataset for use with the \code{bls_map_county()} function.
 #' To access the data directly, issue the command \code{datacounty_map_data)}.
+#' 
+#' @title Dataset for mapping U.S. counties
+#' @description A fortified data set that includes U.S. counties and is suitable for making maps with \code{ggplot2}.
+#' The county FIPS codes and boundary lines from the most recent TIGER release from the U.S. Census Bureau.
 #'
 #' \itemize{
 #'   \item \code{long}: County longitude
@@ -14,10 +18,10 @@
 #' }
 #'
 #' @docType data
-#' @keywords internal
+#' @keywords datasets
 #' 
 #' @name county_map_data
 #' @usage data(county_map_data)
 #' @note Last updated 2016-05-26
 #' @format A data frame with 206,500 rows and 7 variables
-NULL
+"county_map_data"